refactor: Calendar 폴더명 수정
